refactor: categories subfolder
