+++ plugins/announcements/src/components/categories/CategoriesPage/CategoriesPage.tsx
@@ -11,7 +11,7 @@ import {
 import { useApi } from '@backstage/core-plugin-api';
 import { Button, makeStyles } from '@material-ui/core';
 import AddIcon from '@material-ui/icons/Add';
-import { announcementsApiRef, Category } from '../../api';
+import { announcementsApiRef, Category } from '../../../api';
 import { NewCategoryDialog } from '../NewCategoryDialog';
 
 const useStyles = makeStyles(theme => ({
