+++ plugins/announcements/src/components/categories/CategoriesPage/index.ts
