+++ plugins/announcements/src/components/categories/NewCategoryDialog/index.ts
