+++ plugins/announcements/src/components/categories/index.ts
@@ -1,0 +1,2 @@
+export * from './CategoriesPage';
+export * from './NewCategoryDialog';
