+++ plugins/announcements/src/components/homeComponents/index.ts
@@ -1,0 +1,1 @@
+export * from './AnnouncementsCard';
